stock_updater_v1/requirements.txt
# Stock Updater V1.0 依赖

# 数据获取
akshare>=1.12.0
pandas>=2.0.0

# 数据处理
numpy>=1.24.0

# 存储
# sqlite3 (内置)

# 网络请求
requests>=2.28.0

# 调度
schedule>=1.2.0

# YAML 配置
pyyaml>=6.0

# 日志
# logging (内置)
